libsepol/cil: support parallel neverallow checks
Add support to check for neverallow assertions in parallel.  Since the
policy is not modified during those checks there needs to be no
extensive locking.

The checks are run by semodule(8) if the semanage.conf(5) setting
expand-check is set.

Implement the parallel procedure via OpenMP, to minimize code changes
and to allow users a simple way to enable or disable the functionality
at build time.  Currently the support opt-in.

Example benchmark:

    Benchmark 1: ./test_load_wip.sh (serial)
      Time (mean ± σ):      3.485 s ±  0.019 s    [User: 0.003 s, System: 0.002 s]
      Range (min … max):    3.455 s …  3.501 s    5 runs

    Benchmark 1: ./test_load_wip.sh (parallel)
      Time (mean ± σ):      2.443 s ±  0.035 s    [User: 0.004 s, System: 0.001 s]
      Range (min … max):    2.411 s …  2.500 s    5 runs

Signed-off-by: Christian Göttsche <cgzones@googlemail.com>