libsepol: optional data destruction in hashtab_destroy()
Support the destruction of the hashtable entries via an optional
callback in hashtab_destroy(), to avoid iterating the hashtable twice in
common use cases, one time for the entry destruction via hashtab_map()
and a second time via hashtab_destroy() to free the hashtable itself.

Also convert all the destroy callbacks to return void instead of the
needless value of 0.

Signed-off-by: Christian Göttsche <cgzones@googlemail.com>